Add files via upload
This is code for a self balancing skateboard. The hardware associated with this code is two electric motors, a sabertooth motor controller, an accelerator/gyroscope, two 24 V batteries, an arduino Uno, and a voltage transformer. The grapher2.pde graphs the output of the accelerometer so we can see how the noise reduction algorithm is working.